requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
colorama>=0.4.6
argparse>=1.4.0
